TK34792: Start refactoring auth service clients:
1.  Move openstack auth and rackspace auth 1.1 service clients into single maven module.

2.  Refactor so more of the common token code in a single place.

3.  Make RS Auth 1.1 service client implement an interface and update filter component to call against that interface.

4.  Make common Token and Group POJO's that get returned by both OpenStack and Rackspace Auth 1.1 service client.

5.  Update filter components accordingly.
