+++ project-set/components/client-auth/pom.xml
@@ -27,13 +27,18 @@
 
         <dependency>
             <groupId>com.rackspace.cloud.services.clients</groupId>
-            <artifactId>rackspace-auth-v1.1</artifactId>
+            <artifactId>auth</artifactId>
         </dependency>
 
-        <dependency>
-            <groupId>com.rackspace.cloud.services.clients</groupId>
-            <artifactId>openstack-ids</artifactId>
-        </dependency>
+        <!--<dependency>-->
+            <!--<groupId>com.rackspace.cloud.services.clients</groupId>-->
+            <!--<artifactId>rackspace-auth-v1.1</artifactId>-->
+        <!--</dependency>-->
+
+        <!--<dependency>-->
+            <!--<groupId>com.rackspace.cloud.services.clients</groupId>-->
+            <!--<artifactId>openstack-ids</artifactId>-->
+        <!--</dependency>-->
 
         <dependency>
             <groupId>net.sf.ehcache</groupId>
