+++ project-set/components/client-authorization/pom.xml
@@ -1,5 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
 
     <modelVersion>4.0.0</modelVersion>
 
@@ -30,7 +31,7 @@
 
         <dependency>
             <groupId>com.rackspace.cloud.services.clients</groupId>
-            <artifactId>openstack-ids</artifactId>
+            <artifactId>auth</artifactId>
         </dependency>
 
         <dependency>
@@ -75,7 +76,7 @@
 
                 </configuration>
             </plugin>
-            
+
         </plugins>
     </build>
 </project>
