+++ project-set/components/content-identity-auth-1.1/pom.xml
@@ -1,61 +1,62 @@
 <?xml version="1.0" encoding="UTF-8"?>
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-   <modelVersion>4.0.0</modelVersion>
+<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+    <modelVersion>4.0.0</modelVersion>
 
-   <parent>
-      <groupId>com.rackspace.papi.components</groupId>
-      <artifactId>components-support</artifactId>
-      <version>2.0.1-SNAPSHOT</version>
-   </parent>
+    <parent>
+        <groupId>com.rackspace.papi.components</groupId>
+        <artifactId>components-support</artifactId>
+        <version>2.0.1-SNAPSHOT</version>
+    </parent>
 
-   <groupId>com.rackspace.papi.components</groupId>
-   <artifactId>content-identity-auth-1.1</artifactId>
+    <groupId>com.rackspace.papi.components</groupId>
+    <artifactId>content-identity-auth-1.1</artifactId>
 
-   <name>Repose Components - Client Identity By Content - Auth 1.1</name>
-   <description>TODO</description>
+    <name>Repose Components - Client Identity By Content - Auth 1.1</name>
+    <description>TODO</description>
 
-   <packaging>jar</packaging>
-    
-   <dependencies>
-      <dependency>
-         <groupId>com.rackspace.papi.core</groupId>
-         <artifactId>core-lib</artifactId>
-      </dependency>
+    <packaging>jar</packaging>
 
-      <dependency>
-         <groupId>com.rackspace.papi.commons</groupId>
-         <artifactId>configuration</artifactId>
-      </dependency>
+    <dependencies>
+        <dependency>
+            <groupId>com.rackspace.papi.core</groupId>
+            <artifactId>core-lib</artifactId>
+        </dependency>
 
-      <dependency>
-         <groupId>com.rackspace.papi.commons</groupId>
-         <artifactId>utilities</artifactId>
-      </dependency>
-      
-      <dependency>
-         <groupId>com.rackspace.cloud.services.clients</groupId>
-         <artifactId>rackspace-auth-v1.1</artifactId>
-      </dependency>
-      
-      <dependency>
-         <groupId>org.codehaus.jackson</groupId>
-         <artifactId>jackson-mapper-lgpl</artifactId>
-         <version>1.9.4</version>
-         <scope>test</scope>
-      </dependency>
-      
-      <dependency>
-         <groupId>org.codehaus.jackson</groupId>
-         <artifactId>jackson-xc</artifactId>
-         <version>1.9.4</version>
-         <scope>test</scope>
-      </dependency>
+        <dependency>
+            <groupId>com.rackspace.papi.commons</groupId>
+            <artifactId>configuration</artifactId>
+        </dependency>
 
-   </dependencies>
+        <dependency>
+            <groupId>com.rackspace.papi.commons</groupId>
+            <artifactId>utilities</artifactId>
+        </dependency>
 
-   <build>
-      <plugins>
-          <plugin>
+        <dependency>
+            <groupId>com.rackspace.papi.external</groupId>
+            <artifactId>rs-auth-schemas</artifactId>
+        </dependency>
+
+        <dependency>
+            <groupId>org.codehaus.jackson</groupId>
+            <artifactId>jackson-mapper-lgpl</artifactId>
+            <version>1.9.4</version>
+            <scope>test</scope>
+        </dependency>
+
+        <dependency>
+            <groupId>org.codehaus.jackson</groupId>
+            <artifactId>jackson-xc</artifactId>
+            <version>1.9.4</version>
+            <scope>test</scope>
+        </dependency>
+
+    </dependencies>
+
+    <build>
+        <plugins>
+            <plugin>
                 <groupId>org.jvnet.jaxb2.maven2</groupId>
                 <artifactId>maven-jaxb2-plugin</artifactId>
 
@@ -75,7 +76,7 @@
 
                 </configuration>
             </plugin>
-                   
-      </plugins>
-   </build>
+
+        </plugins>
+    </build>
 </project>
