+++ project-set/components/pom.xml
@@ -1,7 +1,8 @@
 <?xml version="1.0" encoding="UTF-8"?>
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <modelVersion>4.0.0</modelVersion>
-    
+
     <parent>
         <groupId>com.rackspace.papi</groupId>
         <artifactId>profile-support</artifactId>
@@ -38,7 +39,7 @@
         <module>header-normalization</module>
         <module>uri-normalization</module>
     </modules>
-    
+
     <dependencyManagement>
         <dependencies>
             <dependency>
@@ -46,7 +47,7 @@
                 <artifactId>echo</artifactId>
                 <version>${project.version}</version>
             </dependency>
-            
+
             <dependency>
                 <groupId>com.rackspace.papi.core</groupId>
                 <artifactId>core-lib</artifactId>
@@ -64,37 +65,55 @@
                 <artifactId>utilities</artifactId>
                 <version>${project.version}</version>
             </dependency>
-            
+
             <dependency>
                 <groupId>com.rackspace.papi.commons</groupId>
                 <artifactId>jetty-container</artifactId>
                 <version>${project.version}</version>
             </dependency>
-            
+
             <dependency>
                 <groupId>com.rackspace.papi.commons</groupId>
                 <artifactId>classloader</artifactId>
                 <version>${project.version}</version>
             </dependency>
-            
+
             <dependency>
                 <groupId>com.rackspace.cloud.services.clients</groupId>
-                <artifactId>rackspace-auth-v1.1</artifactId>
+                <artifactId>auth</artifactId>
                 <version>${project.version}</version>
             </dependency>
-            
+
             <dependency>
-                <groupId>com.rackspace.cloud.services.clients</groupId>
-                <artifactId>openstack-ids</artifactId>
+                <groupId>com.rackspace.papi.external</groupId>
+                <artifactId>rs-auth-schemas</artifactId>
                 <version>${project.version}</version>
             </dependency>
-            
+
+            <dependency>
+                <groupId>com.rackspace.papi.external</groupId>
+                <artifactId>os-auth-schemas</artifactId>
+                <version>${project.version}</version>
+            </dependency>
+
+            <!--<dependency>-->
+            <!--<groupId>com.rackspace.cloud.services.clients</groupId>-->
+            <!--<artifactId>rackspace-auth-v1.1</artifactId>-->
+            <!--<version>${project.version}</version>-->
+            <!--</dependency>-->
+
+            <!--<dependency>-->
+            <!--<groupId>com.rackspace.cloud.services.clients</groupId>-->
+            <!--<artifactId>openstack-ids</artifactId>-->
+            <!--<version>${project.version}</version>-->
+            <!--</dependency>-->
+
             <dependency>
                 <groupId>org.apache.httpcomponents</groupId>
                 <artifactId>httpclient</artifactId>
                 <version>4.1.2</version>
             </dependency>
-            
+
             <dependency>
                 <groupId>log4j</groupId>
                 <artifactId>log4j</artifactId>
@@ -112,7 +131,7 @@
                 <artifactId>ehcache-core</artifactId>
                 <version>2.4.3</version>
             </dependency>
-            
+
             <dependency>
                 <groupId>com.rackspace.papi.components</groupId>
                 <artifactId>default-router</artifactId>
