+++ project-set/external/service-clients/auth/pom.xml
@@ -1,0 +1,55 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project xmlns="http://maven.apache.org/POM/4.0.0"
+         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+    <modelVersion>4.0.0</modelVersion>
+
+    <parent>
+        <artifactId>service-client-support</artifactId>
+        <groupId>com.rackspace.papi.external.clients</groupId>
+        <version>2.0.1-SNAPSHOT</version>
+    </parent>
+
+    <groupId>com.rackspace.cloud.services.clients</groupId>
+    <artifactId>auth</artifactId>
+
+    <name>Repose Service Clients - Auth Service</name>
+
+    <packaging>jar</packaging>
+
+    <dependencies>
+
+        <!-- Due to an issue between javax.mail and another java dependency, this needs to be here for junit to work -->
+        <dependency>
+            <groupId>javax.mail</groupId>
+            <artifactId>mail</artifactId>
+        </dependency>
+
+        <dependency>
+            <groupId>com.rackspace.papi.external</groupId>
+            <artifactId>os-auth-schemas</artifactId>
+            <version>${project.version}</version>
+        </dependency>
+
+        <dependency>
+            <groupId>com.rackspace.papi.external</groupId>
+            <artifactId>rs-auth-schemas</artifactId>
+            <version>${project.version}</version>
+        </dependency>
+
+        <dependency>
+            <groupId>com.rackspace.papi.commons</groupId>
+            <artifactId>utilities</artifactId>
+        </dependency>
+
+        <dependency>
+            <groupId>com.sun.jersey</groupId>
+            <artifactId>jersey-client</artifactId>
+        </dependency>
+
+        <dependency>
+            <groupId>com.sun.jersey</groupId>
+            <artifactId>jersey-server</artifactId>
+        </dependency>
+    </dependencies>
+</project>
