+++ project-set/external/service-clients/auth/src/main/java/com/rackspace/auth/AuthGroup.java
@@ -1,0 +1,10 @@
+package com.rackspace.auth;
+
+/**
+ * @author fran
+ */
+public interface AuthGroup {
+   String getId();
+   String getName();
+   String getDescription();
+}
