+++ project-set/external/service-clients/auth/src/main/java/com/rackspace/auth/rackspace/AuthenticationServiceClientFactory.java
@@ -1,0 +1,15 @@
+package com.rackspace.auth.rackspace;
+
+import com.sun.jersey.api.client.filter.HTTPBasicAuthFilter;
+import com.rackspace.papi.commons.util.http.ServiceClient;
+
+/**
+ * @author fran
+ */
+public class AuthenticationServiceClientFactory {
+   public AuthenticationServiceClient buildAuthServiceClient(String targetHostUri, String username, String password) {
+      HTTPBasicAuthFilter authFilter = new HTTPBasicAuthFilter(username, password);
+
+      return new AuthenticationServiceClient(targetHostUri, new ResponseUnmarshaller(), new ServiceClient(authFilter));
+   }
+}
