--- project-set/external/service-clients/openstack-ids/src/main/java/com/rackspace/auth/openstack/ids/AuthServiceException.java
@@ -1,34 +1,0 @@
-package com.rackspace.auth.openstack.ids;
-
-/*
- *  Copyright 2010 Rackspace.
- *
- *  Licensed under the Apache License, Version 2.0 (the "License");
- *  you may not use this file except in compliance with the License.
- *  You may obtain a copy of the License at
- *
- *       http://www.apache.org/licenses/LICENSE-2.0
- *
- *  Unless required by applicable law or agreed to in writing, software
- *  distributed under the License is distributed on an "AS IS" BASIS,
- *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- *  See the License for the specific language governing permissions and
- *  limitations under the License.
- *  under the License.
- */
-
-
-/**
- *
- * @author jhopper
- */
-public class AuthServiceException extends RuntimeException {
-
-    public AuthServiceException(String message, Throwable cause) {
-        super(message, cause);
-    }
-
-    public AuthServiceException(String message) {
-        super(message);
-    }
-}
