--- project-set/external/service-clients/rackspace-auth-v1.1/pom.xml
@@ -1,48 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-    <modelVersion>4.0.0</modelVersion>
-
-    <parent>
-        <groupId>com.rackspace.papi.external.clients</groupId>
-        <artifactId>service-client-support</artifactId>
-        <version>2.0.1-SNAPSHOT</version>
-    </parent>
-
-    <groupId>com.rackspace.cloud.services.clients</groupId>
-    <artifactId>rackspace-auth-v1.1</artifactId>
-
-    <name>Repose Service Clients - Rackspace Auth v1.1</name>
-
-    <packaging>jar</packaging>
-
-    <dependencies>
-        <dependency>
-            <groupId>net.sf.ehcache</groupId>
-            <artifactId>ehcache-core</artifactId>
-        </dependency>
-
-        <dependency>
-            <groupId>com.rackspace.papi.external</groupId>
-            <artifactId>rs-auth-schemas</artifactId>
-            <version>${project.version}</version>
-        </dependency>
-
-        <dependency>
-            <groupId>com.rackspace.papi.commons</groupId>
-            <artifactId>utilities</artifactId>
-        </dependency>
-
-        <dependency>
-            <groupId>com.sun.jersey</groupId>
-            <artifactId>jersey-client</artifactId>
-        </dependency>
-
-        <dependency>
-            <groupId>com.sun.jersey</groupId>
-            <artifactId>jersey-server</artifactId>
-        </dependency>
-    </dependencies>
-
-</project>
-
