--- project-set/external/service-clients/rackspace-auth-v1.1/src/main/java/com/rackspace/auth/v1_1/ResponseUnmarshaller.java
@@ -1,104 +1,0 @@
-package com.rackspace.auth.v1_1;
-
-import com.rackspace.papi.commons.util.pooling.ConstructionStrategy;
-import com.rackspace.papi.commons.util.pooling.GenericBlockingResourcePool;
-import com.rackspace.papi.commons.util.pooling.ResourceConstructionException;
-import com.rackspace.papi.commons.util.pooling.ResourceContextException;
-import com.rackspacecloud.docs.auth.api.v1.FullToken;
-import com.rackspace.papi.commons.util.pooling.Pool;
-import com.rackspace.papi.commons.util.pooling.ResourceContext;
-
-import javax.xml.bind.JAXBContext;
-import javax.xml.bind.JAXBElement;
-import javax.xml.bind.JAXBException;
-import javax.xml.bind.Unmarshaller;
-import java.io.InputStream;
-import java.io.InputStreamReader;
-import java.io.Reader;
-import java.io.StringReader;
-
-/**
- * @author fran
- */
-public class ResponseUnmarshaller {
-
-   private final JAXBContext jaxbContext;
-   private final Pool<Unmarshaller> pool;
-
-   public ResponseUnmarshaller() {
-      try {
-         jaxbContext = JAXBContext.newInstance(com.rackspacecloud.docs.auth.api.v1.ObjectFactory.class);
-         pool = new GenericBlockingResourcePool<Unmarshaller>(new ConstructionStrategy<Unmarshaller>() {
-
-            @Override
-            public Unmarshaller construct() throws ResourceConstructionException {
-               try {
-                  return jaxbContext.createUnmarshaller();
-               } catch (JAXBException ex) {
-                  throw new ResourceConstructionException("Unable to build jaxb unmarhshaller", ex);
-               }
-            }
-         });
-      } catch (JAXBException jaxbe) {
-         throw new AuthServiceException(
-                 "Possible deployment problem! Unable to build JAXB Context for Auth v1.1 schema types. Reason: "
-                 + jaxbe.getMessage(), jaxbe);
-      }
-   }
-
-   /*
-    * public <T> T unmarshall(final GetMethod method, final Class<T> expectedType) { return pool.use(new
-    * ResourceContext<Unmarshaller, T>() {
-    *
-    * @Override public T perform(Unmarshaller resource) throws ResourceContextException { try { final Object o =
-    * resource.unmarshal(new StringReader(method.getResponseBodyAsString()));
-    *
-    * if (o instanceof JAXBElement && ((JAXBElement) o).getDeclaredType().equals(expectedType)) { return
-    * ((JAXBElement<T>) o).getValue(); } else if (o instanceof FullToken) { return expectedType.cast(o); } else { throw
-    * new AuthServiceException("Failed to unmarshall response body. Unexpected element encountered. Body output is in
-    * debug.");
-    *
-    * }
-    * } catch (IOException ioe) { throw new AuthServiceException("Failed to get response body from response.", ioe); }
-    * catch (JAXBException jaxbe) { throw new AuthServiceException("Failed to unmarshall response body. Body output is
-    * in debug. Reason: " + jaxbe.getMessage(), jaxbe); } } }); }
-    *
-    */
-   private static class UnmarshallerContext<T> implements ResourceContext<Unmarshaller, T> {
-
-      private final Reader reader;
-      private final Class<T> expectedType;
-
-      private UnmarshallerContext(final Reader reader, final Class<T> expectedType) {
-         this.reader = reader;
-         this.expectedType = expectedType;
-      }
-
-      @Override
-      public T perform(Unmarshaller resource) throws ResourceContextException {
-         try {
-            final Object o = resource.unmarshal(reader);
-
-            if (o instanceof JAXBElement && ((JAXBElement) o).getDeclaredType().equals(expectedType)) {
-               return ((JAXBElement<T>) o).getValue();
-            } else if (o instanceof FullToken) {
-               return expectedType.cast(o);
-            } else {
-               throw new AuthServiceException("Failed to unmarshall response body. Unexpected element encountered. Body output is in debug.");
-
-            }
-         } catch (JAXBException jaxbe) {
-            throw new AuthServiceException("Failed to unmarshall response body. Body output is in debug. Reason: "
-                    + jaxbe.getMessage(), jaxbe);
-         }
-      }
-   }
-
-   public <T> T unmarshall(final String data, final Class<T> expectedType) {
-      return pool.use(new UnmarshallerContext<T>(new StringReader(data), expectedType));
-   }
-
-   public <T> T unmarshall(final InputStream data, final Class<T> expectedType) {
-      return pool.use(new UnmarshallerContext<T>(new InputStreamReader(data), expectedType));
-   }
-}
